Added the ability to define a max weight per package.
The max weight per package is useful if you want to

a) limit package weights so they don't go above carrier limits like UPS's 150lb limit.
b) have a company rule that splits boxes after a certain weight.

It is implemented to respect the countries max weight if that happens to be lower.

Signed-off-by: Nathan Lowrie <nate@finelineautomation.com>