Added the ability to create product packages
Product Packages are a way for products that ship in there own packages to be defined properly for shipping estimates.  It is extremely useful for certain cases like a 200 lb product that ships in 5x 40lb boxes because 200lbs throws a UPS ship error.  This commit is only the entering of product packages through the admin interface.

Signed-off-by: Nathan Lowrie <nate@finelineautomation.com>