Added Product Packages to calculator.
The calculator now includes product packages in it's base calculation.  The user can now enter packages for a particular product and have those packages be used in the base freight calculation.  The calculator checks the products for over weight limit packages and will raise an error if the packages weigh too much.

Signed-off-by: Nathan Lowrie <nate@finelineautomation.com>