Fixes #76
This patch fixes an issue where we were adding in weights for other shipments coming from other stock locations. I think this properly splits out the relevant items in the shipment. I'm not clear if this has broken anything in the calcs that look at max weights and multiple packages. From what I have learned, I don't think a lot of that code specific to multiple packages could have worked as is...